requirements.txt
appnope==0.1.0
argcomplete==1.11.1
argh==0.26.2
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
awscli==1.18.2
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.8.2
bleach==3.1.0
blessed==1.17.2
boto3==1.12.2
botocore==1.15.2
cached-property==1.5.1
cement==3.0.4
certifi==2019.11.28
cfn-flip==1.2.2
chardet==3.0.4
click==7.0
colorama==0.4.3
compass-stylesheets==1.0.2
cycler==0.10.0
data-wizard==1.2.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.3
django-adaptors==0.2.5
django-cors-headers==3.2.1
django-filter==2.2.0
django-letsencrypt==3.0.1
django-modelcluster==5.0.1
django-mustache==1.1.2
django-redis-cache==2.1.0
django-sendgrid-v5==0.8.1
django-sslserver==0.22
django-storages==1.9.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
djangorestframework-csv==2.1.0
docker==4.2.0
docker-compose==1.25.4
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
draftjs-exporter==3.0.0
durationpy==0.5
entrypoints==0.3
et-xmlfile==1.0.1
FormEncode==1.3.1
future==0.18.2
gunicorn==20.0.4
hjson==3.0.1
html-json-forms==1.1.1
html5lib==1.0.1
idna==2.9
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
linecache2==1.0.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
natural-keys==1.6.0
nbconvert==5.6.1
nbformat==5.0.4
nose==1.3.7
notebook==6.0.3
numpy==1.18.1
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pathspec==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
placebo==0.9.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pur==5.3.0
py-mini-racer==0.1.18
pyasn1==0.4.8
PyBabeljs==0.0.8
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
pyScss==1.3.5
pystache==0.5.4
python-dateutil==2.8.1
python-http-client==3.2.4
python-slugify==4.0.0
pytz==2019.3
pyxform==1.1.0
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requirejs==0.2.1
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.22.1
scipy==1.4.1
semantic-version==2.8.4
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
sqlparse==0.3.0
style==1.1.6
sympy==1.5.1
tablib==1.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
toml==0.10.0
tornado==6.0.3
tqdm==4.43.0
traceback2==1.4.0
traitlets==4.3.3
troposphere==2.5.3
typed-ast==1.4.1
unicodecsv==0.14.1
Unidecode==1.1.1
unittest2==1.1.0
update==0.0.1
urllib3==1.25.8
virtualenv==20.0.4
wagtail==2.8
wagtail-robots==0.1.2
wagtail-schema.org==2.0.0
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.0
whitenoise==5.0.1
widgetsnbextension==3.5.1
Willow==1.3
wq.app==1.2.0
wq.core==1.2.0
wq.db==1.2.0
wq.io==1.1.0
wq.start==1.2.0
wrapt==1.12.0
wsgi-request-logger==0.4.6
xlrd==1.2.0
xlsconv==1.1.0
xlwt==1.3.0
